+++ obsidian/tests/test_optimizer_SOO.py
@@ -7,7 +7,6 @@ from obsidian.optimizer import BayesianOptimizer
 from obsidian.constraints import OutConstraint_Blank, InConstraint_Generic, InConstraint_ConstantDim, OutConstraint_L1
 from obsidian.objectives import Identity_Objective, Feature_Objective, Objective_Sequence, Utopian_Distance, Bounded_Target
 from obsidian.tests.utils import approx_equal
-from obsidian.campaign.explainer import Explainer
 
 import pandas as pd
 import numpy as np
@@ -167,13 +166,6 @@ def test_combo_constraints():
 def test_objective(m_batch, obj):
     X_suggest, eval_suggest = optimizer.suggest(m_batch=m_batch, objective=obj, **test_config)
     df_suggest = pd.concat([X_suggest, eval_suggest], axis=1)
-
-
-@pytest.mark.fast
-def test_explain():
-    model_exp = Explainer(optimizer)
-    model_exp.shap_explain(n=10)
-    df_sens = model_exp.cal_sensitivity(dx=0.01)
 
 
 if __name__ == '__main__':